// src/types/course.types.ts
export interface ICourse {
  id?: string;
  title: string;
  description: string;
  // imageName?: string;
  // instructor?: string;
  price?: number;
  discountedPrice?: number;
  // currency?: string;
  // country?: string;
  // rating?: number;
  // courseUrl?: string;
  // refundable?: boolean;
  // liveOrRecorded?: string;
}